requirements.txt
google_api_python_client==2.111.0
numpy==1.26.4
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
streamlit==1.29.0
streamlit==1.30.0
transformers==4.36.2
transformers==4.37.2
wget==3.2
youtube_transcript_api==0.6.1
